+++ gadgets/mri_core/CropAndCombineGadget.cpp
@@ -22,27 +22,27 @@ process( GadgetContainerMessage<ISMRMRD::ImageHeader>* m1,
     return -1;
   }
 
-  int dimx     = m3->getObjectPtr()->get_size(0);
-  int dimx_old = m2->getObjectPtr()->get_size(0);
+  size_t dimx     = m3->getObjectPtr()->get_size(0);
+  size_t dimx_old = m2->getObjectPtr()->get_size(0);
 
-  int dimy = m3->getObjectPtr()->get_size(1);
-  int dimz = m3->getObjectPtr()->get_size(2);
+  size_t dimy = m3->getObjectPtr()->get_size(1);
+  size_t dimz = m3->getObjectPtr()->get_size(2);
 
-  int channels = m2->getObjectPtr()->get_size(3);
+  size_t channels = m2->getObjectPtr()->get_size(3);
 
   std::complex<float>* d1 = m2->getObjectPtr()->get_data_ptr();
   std::complex<float>* d2 = m3->getObjectPtr()->get_data_ptr();
 
   size_t img_block_old = dimx_old*dimy*dimz;
 
-  for (int z = 0; z < dimz; z++) {
-    for (int y = 0; y < dimy; y++) {
-      for (int x = 0; x < dimx; x++) {
+  for (size_t z = 0; z < dimz; z++) {
+    for (size_t y = 0; y < dimy; y++) {
+      for (size_t x = 0; x < dimx; x++) {
 	float mag = 0;
 	float phase = 0;
 	size_t offset_1 = z*dimy*dimx_old+y*dimx_old+x+((dimx_old-dimx)>>1);
 	size_t offset_2 = z*dimy*dimx+y*dimx+x;
-	for (int c = 0; c < channels; c++) {
+	for (size_t c = 0; c < channels; c++) {
 	  float mag_tmp = norm(d1[offset_1 + c*img_block_old]);
 	  phase += mag_tmp*arg(d1[offset_1 + c*img_block_old]);
 	  mag += mag_tmp;
