+++ toolboxes/gtplus/workflow/gtPlusISMRMRDReconWorkFlowCartesian3DT.h
@@ -221,7 +221,7 @@ bool gtPlusISMRMRDReconWorkFlowCartesian3DT<T>::recon()
 
                 if ( dim == WorkOrderShareDim_ )
                 {
-                    indWorkOrderSharingDim = dims.size()-1;
+                    indWorkOrderSharingDim = (int)(dims.size()-1);
                 }
             }
         }
